compare button, model info
There was a little bit of an issue with the compare button. Basically, it was getting enabled even when the models and datasets all were different. Inclusion of modelIDList in the state along with a little tweak in the terenary operator condition for enabling, fixed the issue. Other than that, removed some ESLint warnings and added the Model Information tab in all the use cases